refactor(test): rename JET to Inference ...
Also: moved the Julia version check to the runtests file.
