--- JET/Project.toml
@@ -1,2 +1,0 @@
-[deps]
-Mantis = "45fd81cf-4570-4cb4-b51d-06adeaefc15d"
